Fix process of adding header image
An image found in the project folder will be given priority over an
image url on the trello card. This is to streamline the process between
making the image and putting in the newsletter.

Instead of having to make the image, upload it, grab the link
and then put in trello. We can bypass this by placing the image you make
into the project folder.